% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvr.R
\name{bootstrap_bvr_pvals}
\alias{bootstrap_bvr_pvals}
\title{Bootstrap p-values of the bivariate residuals of an LCA model.}
\usage{
bootstrap_bvr_pvals(formula, fit_polca, data, R = 200, ...)
}
\arguments{
\item{formula}{A formula to be used as input by poLCA}

\item{fit_polca}{A poLCA model fit object.}

\item{data}{Original data used to fit \code{fit_polca}.}

\item{R}{The number of parametric bootstrap replicates.}

\item{...}{Further arguments to \code{poLCA()}, such as \code{nrep} and \code{nclass}}
}
\value{
Bootstrapped p-values of the BVRs
}
\description{
Bootstrap p-values of the bivariate residuals of an LCA model.
}
