% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvr.R
\name{bvr}
\alias{bvr}
\title{Calculate bivariate residuals in a poLCA model.}
\usage{
bvr(fit, tol = 0.001, rescale_to_df = TRUE)
}
\arguments{
\item{fit}{A poLCA model fit object}

\item{tol}{A tolerance (scalar double) to deal with zero counts}

\item{rescale_to_df}{In bivariate crosstables for \emph{polytomous} variables,
whether to divide the resulting 'chi-square' type value by its
degrees of freedom. Defaults to true.}
}
\description{
Calculate bivariate residuals in a poLCA model.
}
