% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvr.R
\name{get_fitter_and_calculator}
\alias{get_fitter_and_calculator}
\title{Output a function that fits a poLCA model to data and calculates BVRs on it.}
\usage{
get_fitter_and_calculator(f, ...)
}
\arguments{
\item{f}{A formula to be used as input by poLCA}

\item{...}{further arguments to poLCA, such as the number of classes \code{nclass}}
}
\value{
A function to be used as \verb{statistic=} by \code{boot()}
}
\description{
Output a function that fits a poLCA model to data and calculates BVRs on it.
}
