% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvr.R
\name{get_generator}
\alias{get_generator}
\title{Given a poLCA fit object, output a function that generates data from it.}
\usage{
get_generator(fit_polca)
}
\arguments{
\item{fit_polca}{A poLCA model fit object}
}
\value{
A function to be used as \verb{ran.gen=} by \code{boot()}
}
\description{
Given a poLCA fit object, output a function that generates data from it.
}
