% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvr.R
\name{get_fitter_and_calculator}
\alias{get_fitter_and_calculator}
\title{Some convenience function for parametric bootstrapping
Given a formula, and other arguments to poLCA,
output a  function : data → statistics:
the output function fits poLCA to data as requested and
returns the statistics of interest.}
\usage{
get_fitter_and_calculator(f, ...)
}
\arguments{
\item{f}{A formula to be used as input by poLCA}

\item{...}{further arguments to poLCA, such as the number of classes \code{nclass}}
}
\value{
A function to be used as \verb{statistic=} by \code{boot()}
}
\description{
Some convenience function for parametric bootstrapping
Given a formula, and other arguments to poLCA,
output a  function : data → statistics:
the output function fits poLCA to data as requested and
returns the statistics of interest.
}
